thndr_task/pyproject.toml
[tool.poetry]
name = "thndr_task"
version = "0.1.0"
description = "Thndr Backend Task"
authors = ["Ahmed Samir <a.samir9710@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
djangorestframework = "^3.13.1"
paho-mqtt = "^1.6.1"
redis = "^4.3.4"
drf-yasg = "^1.21.3"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
